feat: enable authorization
Implement authorization for creating, deleting and updating a book on the nestjs app.
Include the auth0 token(if present) as a authorization header in every request from the react app.
Tested successfuly with and without the auth token in the graphql playground.
Some frontend UIs are needed to come full circle.